CUIモードの追加
